Change integer types to fixed width

+++ Sources/Wrapper/RNPeerInterface.h
@@ -34,22 +34,22 @@
 /**
  Return the total number of connections we are allowed.
  */
-@property (readonly, nonatomic) unsigned int maximumNumberOfPeers;
+@property (readonly, nonatomic) uint32_t maximumNumberOfPeers;
 
 /**
  Maximum number of incoming connections allowed.
  */
-@property (nonatomic) unsigned short maximumIncomingConnections;
+@property (nonatomic) uint16_t maximumIncomingConnections;
 
 /**
  Returns how many open connections there are at this time.
  */
-@property (readonly, nonatomic) unsigned short numberOfConnections;
+@property (readonly, nonatomic) uint16_t numberOfConnections;
 
 /**
  Return own GUID.
  */
-@property (readonly, nonatomic) unsigned long long myGUID;
+@property (readonly, nonatomic) uint64_t myGUID;
 
 /**
  Returns an array of IP addresses this system has internally.
@@ -87,7 +87,7 @@
  
  @return YES if sturtup succeeded, NO otherwise.
  */
-- (BOOL)startupWithMaxConnectionsAllowed:(unsigned int)maxConnections
+- (BOOL)startupWithMaxConnectionsAllowed:(uint32_t)maxConnections
                        socketDescriptors:(nonnull NSArray *)descriptors
                                    error:(out NSError * __nullable * __nullable)error
 NS_SWIFT_NAME(startup(maxConnections:socketDescriptors:));
@@ -168,7 +168,7 @@ NS_SWIFT_NAME(securityExceptionList(hasAddress:));
  @return YES on successful initiation, NO otherwise.
  */
 - (BOOL)connectToHost:(nonnull NSString *)host
-           remotePort:(unsigned short)remotePort
+           remotePort:(uint16_t)remotePort
              password:(nullable NSData *)password
             publicKey:(nullable RNPublicKey *)publicKey
                 error:(out NSError * __nullable * __nullable)error
@@ -180,7 +180,7 @@ NS_SWIFT_NAME(connect(remoteHost:remotePort:password:publicKey:));
  @param guid             The GUID of the system we are checking to see if we are connected to.
  @param sendNotification YES to send .DisconnectionNotification to the recipient. NO to close it silently.
  */
-- (void)disconnectRemoteGUID:(unsigned long long)guid
+- (void)disconnectRemoteGUID:(uint64_t)guid
                sendNotification:(BOOL)sendNotification
 NS_SWIFT_NAME(disconnect(remoteGUID:notify:));
 
@@ -203,7 +203,7 @@ NS_SWIFT_NAME(disconnect(remoteAddress:notify:));
 
  @return What state the remote system is in.
  */
-- (RNConnectionState)getConnectionStateWithGUID:(unsigned long long)guid
+- (RNConnectionState)getConnectionStateWithGUID:(uint64_t)guid
 NS_SWIFT_NAME(connectionState(remoteGUID:));
 
 /**
@@ -266,7 +266,7 @@ NS_SWIFT_NAME(unban(address:));
  @return YES on success, NO on failure (unknown hostname)
  */
 - (BOOL)pingAddress:(nonnull NSString *)address
-         remotePort:(unsigned short)remotePort
+         remotePort:(uint16_t)remotePort
 onlyReplyOnAcceptingConnections:(BOOL)onlyReplyOnAcceptingConnections
 NS_SWIFT_NAME(ping(address:port:replyOnAcceptingConnections:));
 
@@ -285,7 +285,7 @@ NS_SWIFT_NAME(ping(address:));
 
  @return The last ping time for this system, or -1.
  */
-- (int)getLastPingForGUID:(unsigned long long)guid
+- (int)getLastPingForGUID:(uint64_t)guid
 NS_SWIFT_NAME(getLastPing(forGUID:));
 
 /**
@@ -368,7 +368,7 @@ NS_SWIFT_NAME(send(data:priority:reliability:channel:guid:broadcast:receipt:));
 
  @return Address of the system
  */
-- (nonnull RNSystemAddress *)getSystemAddressFromGUID:(unsigned long long)guid
+- (nonnull RNSystemAddress *)getSystemAddressFromGUID:(uint64_t)guid
 NS_SWIFT_NAME(getAddress(fromGUID:));
 
 /**
@@ -378,7 +378,7 @@ NS_SWIFT_NAME(getAddress(fromGUID:));
 
  @return The GUID of the system
  */
-- (unsigned long long)getGuidFromSystemAddress:(nonnull RNSystemAddress *)address
+- (uint64_t)getGuidFromSystemAddress:(nonnull RNSystemAddress *)address
 NS_SWIFT_NAME(getGUID(fromAddress:));
 
 @end
